implemented testing to address enhancement #1

+++ controllers/Default.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const prismaClient = require("../client")
+const Context = require("../context")
 const { check, validationResult } = require('express-validator');
 var utils = require('../utils/writer.js');
 var Default = require('../service/DefaultService');
@@ -16,12 +16,16 @@ module.exports.addGroup = function addGroup (req, res, next) {
 };
 
 module.exports.addUser = function addUser (req, res, next, body) {
-
-  Default.addUser(body.userEmail, prismaClient)
+  if (Default.userExists(body.userEmail, Context())){
+  Default.addUser(body.userEmail, Context())
     .then(function (response) {
       utils.writeJson(res, response);
     })
     .catch(function (response) {
       utils.writeJson(res, response);
     });
+  }
+  else {
+    utils.writeJson(res, {"error": "user already exists"}, 409)
+  }
 };
